services/gateway/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
pydantic==2.9.2
prometheus-client==0.20.0
httpx==0.27.2
